refactor: Structural output Android models describe
