+++ test_runner/src/main/kotlin/ftl/domain/DescribeAndroidModels.kt
@@ -1,19 +1,21 @@
 package ftl.domain
 
-import flank.common.logLn
 import ftl.api.fetchDeviceModelAndroid
 import ftl.args.AndroidArgs
-import ftl.environment.android.getDescription
+import ftl.presentation.Output
 import ftl.run.exception.FlankConfigurationError
+import ftl.run.exception.FlankGeneralError
 import java.nio.file.Paths
 
-interface DescribeAndroidModels {
+interface DescribeAndroidModels : Output {
     val configPath: String
     val modelId: String
 }
 
 operator fun DescribeAndroidModels.invoke() {
     if (modelId.isBlank()) throw FlankConfigurationError("Argument MODEL_ID must be specified.")
-    // TODO move getDescription() and printing presentation layer during refactor of presentation after #1728
-    logLn(fetchDeviceModelAndroid(AndroidArgs.loadOrDefault(Paths.get(configPath)).project).getDescription(modelId))
+    fetchDeviceModelAndroid(AndroidArgs.loadOrDefault(Paths.get(configPath)).project)
+        .find { it.id == modelId }
+        ?.out()
+        ?: throw FlankGeneralError("ERROR: '$modelId' is not a valid model")
 }
